requirements.txt
google-cloud-texttospeech
requests
beautifulsoup4
html5lib
python-slugify
pydub
